+++ src/assemble.sh
@@ -15,6 +15,7 @@ print_usage_and_exit()
     echo "    $0 $CMD_GAMBATTE"
     echo "    $0 $CMD_MOONEYE_GB"
     echo "    $0 $CMD_RGBDS"
+    echo "    $0 $CMD_DMG_ACID2"
     echo "    $0 $CMD_RELEASE_ZIP <zip-file>"
     exit 1
 }
@@ -152,6 +153,29 @@ build_rgbds()
 
 
 
+build_dmg_acid2()
+{
+    # extract RGBDS artifacts
+    cd "$ARTIFACTS_DIR"
+    untar_all_artifacts
+    PATH="$ARTIFACTS_DIR/rgbds:$PATH"
+
+    ARTIFACT_NAME=dmg-acid2
+    ARTIFACT=$(mkdir_artifact $ARTIFACT_NAME)
+
+    REPO_DMG_ACID2=$(mktemp -d)
+    cd "$REPO_DMG_ACID2"
+    git clone --recurse-submodules https://github.com/mattcurrie/dmg-acid2 .
+    git checkout 8a98ce731f96dde032ffb22ec36dc985d78fdb18
+    make
+
+    cp build/dmg-acid2.gb "$ARTIFACT"
+
+    tar_rm_artifact $ARTIFACT_NAME
+}
+
+
+
 create_release_zip()
 {
     ZIP_FILE=$1
@@ -168,6 +192,7 @@ create_release_zip()
 
     # create new zip file
     untar_all_artifacts
+    rm -rf rgbds
     zip -r "$ZIP_FILE" .
 }
 
@@ -183,6 +208,7 @@ CMD_BLARGG=blargg-roms
 CMD_GAMBATTE=gambatte-roms
 CMD_MOONEYE_GB=mooneye-gb-roms
 CMD_RGBDS=rgbds
+CMD_DMG_ACID2=dmg-acid2
 CMD_RELEASE_ZIP=release-zip
 
 # identify repository directories based on the path of this script
@@ -202,6 +228,7 @@ case ${CMD} in
     ${CMD_GAMBATTE}) assemble_gambatte $@ ;;
     ${CMD_MOONEYE_GB}) assemble_mooneye_gb $@ ;;
     ${CMD_RGBDS}) build_rgbds $@ ;;
+    ${CMD_DMG_ACID2}) build_dmg_acid2 $@ ;;
     ${CMD_RELEASE_ZIP}) create_release_zip $@ ;;
 
     *) print_usage_and_exit ;;